Move Write methods to separate files.
